feat: rename with .web to avoid bundling non required files on native
